[Chore] 원시값 타입 CGFloat으로 변경

+++ Score/Score/Constant/Constants.swift
@@ -87,7 +87,7 @@ enum Constants {
     
     //MARK: - Layout
     
-    enum Layout: Int {
+    enum Layout: CGFloat {
         case horizontal = 24
     }
 }
